previo a contrato

+++ pescolar-web/src/main/java/sv/gob/mined/pescolar/model/ContratosOrdenesCompra.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
             name = "SP_ADD_RESUMEN_CE_PROCESADO",
             procedureName = "SP_ADD_RESUMEN_CE_PROCESADO",
             parameters = {
-                @StoredProcedureParameter(name = "P_ID_DET_PROCESO_ADQ", mode = ParameterMode.IN, type = Integer.class),
+                @StoredProcedureParameter(name = "P_ID_DET_PROCESO_ADQ", mode = ParameterMode.IN, type = Long.class),
                 @StoredProcedureParameter(name = "P_CODIGO_ENTIDAD", mode = ParameterMode.IN, type = String.class),
                 @StoredProcedureParameter(name = "P_CODIGO_DEPARTAMENTO", mode = ParameterMode.IN, type = String.class),
                 @StoredProcedureParameter(name = "P_CODIGO_MUNICIPIO", mode = ParameterMode.IN, type = String.class)
@@ -24,7 +24,7 @@ import java.time.LocalDateTime;
             name = "SP_ADD_RESUMEN_ADJ_EMP",
             procedureName = "SP_ADD_RESUMEN_ADJ_EMP",
             parameters = {
-                @StoredProcedureParameter(name = "P_ID_DET_PROCESO_ADQ", mode = ParameterMode.IN, type = Integer.class),
+                @StoredProcedureParameter(name = "P_ID_DET_PROCESO_ADQ", mode = ParameterMode.IN, type = Long.class),
                 @StoredProcedureParameter(name = "P_ID_EMPRESA", mode = ParameterMode.IN, type = String.class),
                 @StoredProcedureParameter(name = "P_ID_TIPO_EMPRESA", mode = ParameterMode.IN, type = String.class),
                 @StoredProcedureParameter(name = "P_CODIGO_DEPARTAMENTO", mode = ParameterMode.IN, type = String.class),
